abstaraction
