+++ banditpylib/learners/collaborative_learner/__init__.py
@@ -1,5 +1,7 @@
 from .utils import *
 from .lilucb_heur_collaborative import *
+from .colaborative_learning import *
 
-__all__ = ['CollaborativeLearner', 'CollaborativeAgent',
-    'CollaborativeMaster']
+__all__ = ['CollaborativeBAILearner', 'CollaborativeBAIAgent',
+    'CollaborativeBAIMaster', 'LilUCBHeuristicCollaborativeBAIAgent',
+    'LilUCBHeuristicCollaborativeBAIMaster']
