+++ banditpylib/learners/collaborative_learner/lilucb_heur_collaborative_test.py
@@ -15,17 +15,17 @@ class TestLilUCBHeuristicCollaborative:
 
     while True:
       actions = learner.actions()
-      if not actions.arm_pulls_pairs:
+      if not actions.arm_pulls:
         break
 
       feedback = Feedback()
-      for arm_pulls_pair in actions.arm_pulls_pairs:
-        arm_rewards_pair = feedback.arm_rewards_pairs.add()
-        arm_rewards_pair.arm.id = arm_pulls_pair.arm.id
-        arm_rewards_pair.rewards.extend(
+      for arm_pull in actions.arm_pulls:
+        arm_feedback = feedback.arm_feedbacks.add()
+        arm_feedback.arm.id = arm_pull.arm.id
+        arm_feedback.rewards.extend(
             list(
-                np.random.normal(arm_pulls_pair.arm.id / arm_num, 1,
-                                 arm_pulls_pair.pulls)))
+                np.random.normal(arm_pull.arm.id / arm_num, 1,
+                                 arm_pull.times)))
       learner.update(feedback)
 
     assert learner.best_arm in list(range(arm_num))
